requirements.txt
Django==1.11.10
django-allauth==0.35.0
google-api-python-client==1.6.5
ipython==6.2.1
ipython-genutils==0.2.0
jsonpickle==0.9.6
pytz==2018.3
django-multi-email-field==0.5.1
django-widget-tweaks==1.4.2
phonenumbers==8.9.4
oauth2client==4.1.2
absl-py
